doc: get doc compiling again
MethodSetCache moved to typeutil; thanks to tiago4orion for the heads up.

However, this should just be deleted once 1.5 lands as it is replaced by the
new go doc, although it does have a couple of features not present there.

Meanwhile, this change brings up an issue in the 1.5beta, which is that you
can't build this program there because MethodSetCache is not present in
the core.